+++ v2/src/views/ViewSettings.vue
@@ -18,6 +18,10 @@ import ModuleAddressNames from '@/modules/settings/ModuleAddressNames.vue'
 import ModuleSettings from '@/modules/settings/ModuleSettings.vue'
 import { useAppViewGrid } from '@core/composables/AppViewGrid/AppViewGrid.composable'
 import { onMounted } from 'vue'
+import { VIEW_TAGS } from '@core/helper/tags'
+import { usePageMeta } from '@core/composables/use-page-meta/use-page-meta.composable'
+
+usePageMeta(null, VIEW_TAGS.SETTINGS)
 
 const { columnPadding, rowMargin } = useAppViewGrid()