+++ v2/src/views/ViewUncleDetails.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script setup lang="ts">
-import { reactive, computed, onMounted } from 'vue'
+import { reactive, computed, onMounted, toRefs } from 'vue'
 import AppError from '@core/components/AppError.vue'
 import AppAdButtonsLarge from '@/core/components/AppAdButtonsLarge.vue'
 import UncleDetails from '@module/uncles/ModuleUncleDetails.vue'
@@ -26,6 +26,8 @@ import { eth } from '@core/helper'
 import { ErrorMessageUncle } from '@module/uncles/models/ErrorMessagesForUncle'
 import { useAppViewGrid } from '@core/composables/AppViewGrid/AppViewGrid.composable'
 import { useI18n } from 'vue-i18n'
+import { VIEW_TAGS } from '@core/helper/tags'
+import { usePageMeta } from '@core/composables/use-page-meta/use-page-meta.composable'
 
 const { t } = useI18n()
 const { columnPadding, rowMargin } = useAppViewGrid()
@@ -41,8 +43,13 @@ const state: ComponentState = reactive({
 })
 
 const props = defineProps({
-    uncleRef: String
+    uncleRef: {
+        type: String,
+        required: true
+    }
 })
+const { uncleRef } = toRefs(props)
+usePageMeta(uncleRef, VIEW_TAGS.UNCLE)
 
 const isValid = computed<boolean>(() => {
     return eth.isValidHash(props.uncleRef)